+++ flink-connectors/flink-connector-hive/src/test/java/org/apache/flink/table/catalog/hive/HiveTestUtils.java
@@ -40,6 +40,13 @@ public class HiveTestUtils {
 		return new GenericHiveMetastoreCatalog(CatalogTestBase.TEST_CATALOG_NAME, getHiveConf());
 	}
 
+	/**
+	 * Create a HiveCatalog with an embedded Hive Metastore.
+	 */
+	public static HiveCatalog createHiveCatalog() throws IOException {
+		return new HiveCatalog(CatalogTestBase.TEST_CATALOG_NAME, getHiveConf());
+	}
+
 	private static HiveConf getHiveConf() throws IOException {
 		ClassLoader classLoader = new HiveTestUtils().getClass().getClassLoader();
 		HiveConf.setHiveSiteLocation(classLoader.getResource(HIVE_SITE_XML));
